update test.py. renamed test_examples.py. compares all output files
